telegram_bot/requirements.txt
colorlog
requests
isort
pylint
python-telegram-bot
python-telegram-bot[job-queue]
ollama
python-dotenv